requirements.txt
easydict==1.13
opencv-python==4.9.0.80
moviepy==1.0.3
flow-vis==0.1
matplotlib==3.8.3
einops==0.7.0
timm==1.0.15
scikit-image==0.22.0
scikit-learn==1.4.1.post1
cupy-cuda11x
accelerate
yt-dlp
pandas
bitsandbytes
diffusers==0.36.0
transformers==4.57.3
hf_transfer>=0.1.8
peft>=0.12.0
decord>=0.6.0
wandb
torchao==0.9.0
sentencepiece>=0.2.0
imageio-ffmpeg>=0.5.1
numpy>=1.26.4
deepspeed
gradio
setuptools<82

# HYPIR https://github.com/XPixelGroup/HYPIR/blob/main/requirements.txt
open_clip_torch==2.31.0
vision-aided-loss==0.1.0
lpips==0.1.4
